# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "York academy",
        "description": "York academy",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/"
        },
        {
            "url": "/api/"
        }
    ],
    "paths": {
...
        "/course_ads/{id}": {
            "get": {
                "tags": [
                    "Course Ads"
                ],
                "summary": "Get course ad by ID",
                "description": "Retrieve details of a specific course ad by its ID",
                "operationId": "getCourseAdsById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the course ad",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course ad details retrieved successfully"
                    },
                    "404": {
                        "description": "Course ad not found"
                    }
                }
            }
        },
...
        "/registration": {
            "post": {
                "tags": [
                    "Registrations"
                ],
                "summary": "Create a new registration",
                "description": "Store a new registration for a course",
                "operationId": "createRegistration",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "course_id",
                                    "full_name",
                                    "phone",
                                    "email",
                                    "gender",
                                    "address",
                                    "notes"
                                ],
                                "properties": {
                                    "course_id": {
                                        "description": "ID of the course",
                                        "type": "integer"
                                    },
                                    "full_name": {
                                        "description": "Full name of the registrant",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "Phone number of the registrant",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Email address of the registrant",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "gender": {
                                        "description": "Gender of the registrant",
                                        "type": "string"
                                    },
                                    "address": {
                                        "description": "Address of the registrant",
                                        "type": "string"
                                    },
                                    "notes": {
                                        "description": "Additional notes for the registration",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registration created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Created Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Course not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Course data not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/registration/{id}": {
            "get": {
                "tags": [
                    "Registrations"
                ],
                "summary": "Get details of a specific registration",
                "description": "Retrieve the registration details by ID",
                "operationId": "getRegistrationById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the registration to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registration data retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Retrieved successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Registration"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Registration not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Registration not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/sections/{id}": {
            "get": {
                "tags": [
                    "Sections"
                ],
                "summary": "Get details of a specific section",
                "description": "Retrieve the section details by ID",
                "operationId": "getSectionById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the section to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Section data retrieved successfully"
                    },
                    "404": {
                        "description": "Section not found"
                    }
                }
            }
        },
...
        "/sliders/{id}": {
            "get": {
                "tags": [
                    "Sliders"
                ],
                "summary": "Get details of a specific slider",
                "description": "Retrieve the slider details by ID",
                "operationId": "getSliderById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the slider to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "slider data retrieved successfully"
                    },
                    "404": {
                        "description": "slider not found"
                    }
                }
            }
        },
...
        "/training_plan/latestPlan": {
            "get": {
                "tags": [
                    "Training Plans"
                ],
                "summary": "Get the latest training plan for the current year",
                "description": "Retrieve the most recent training plan for the current year, if available",
                "operationId": "getLatestTrainingPlan",
                "responses": {
                    "200": {
                        "description": "Latest training plan retrieved successfully"
                    },
                    "404": {
                        "description": "Training plan not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Registration": {
                "description": "Registration model",
                "properties": {
                    "id": {
                        "description": "ID of the registration",
                        "type": "integer"
                    },
                    "course_id": {
                        "description": "ID of the course",
                        "type": "integer"
                    },
                    "full_name": {
                        "description": "Full name of the registrant",
                        "type": "string"
                    },
                    "phone": {
                        "description": "Phone number of the registrant",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email address of the registrant",
                        "type": "string",
                        "format": "email"
                    },
                    "gender": {
                        "description": "Gender of the registrant",
                        "type": "string"
                    },
                    "address": {
                        "description": "Address of the registrant",
                        "type": "string"
                    },
                    "notes": {
                        "description": "Additional notes for the registration",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
...
    },
...
}